Moves static pages to their own folder. Removes unecessary imports #81
