+++ src/App.js
@@ -19,16 +19,15 @@ import EquipmentReservations from './components/entities/EquipmentReservations'
 import Membership from './components/Membership'
 import ReservationManagement from './components/ReservationManagement'
 import ReservationsList from './components/ReservationsList'
-import AboutUs from './components/AboutUs'
-import FacilityHoursAndSchedule from './components/FacilityHoursAndSchedule'
-import StayConnected from './components/StayConnected'
+import AboutUs from './components/static_pages/AboutUs'
+import FacilityHoursAndSchedule from './components/static_pages/FacilityHoursAndSchedule'
+import StayConnected from './components/static_pages/StayConnected'
 import EquipmentBundle from './components/entities/EquipmentBundle'
-import AquaticCenter from './components/AquaticCenter'
-import RecreationCenter from './components/RecreationCenter'
-import FitnessCourt from './components/FitnessCourt'
-import OutdoorComplex from './components/OutdoorComplex'
-import PlannedProjects from './components/PlannedProjects'
-import Policies from './components/Policies'
+import AquaticCenter from './components/static_pages/AquaticCenter'
+import RecreationCenter from './components/static_pages/RecreationCenter'
+import FitnessCourt from './components/static_pages/FitnessCourt'
+import OutdoorComplex from './components/static_pages/OutdoorComplex'
+import PlannedProjects from './components/static_pages/PlannedProjects'
 import Notification from "./components/Notification"
