+++ src/components/static_pages/FacilityHoursAndSchedule.js
@@ -1,9 +1,7 @@
 import React from 'react'
 import { makeStyles } from '@material-ui/core/styles'
-import RecCenterVector from '../static/RecCenter.svg'
-import { Button, Container, Table, TableCell, TableHead,
+import { Container, Table, TableCell, TableHead,
   TableRow, TableBody, Paper } from '@material-ui/core'
-import { BrowserRouter as Router, Route, Link } from 'react-router-dom'
 
 const useStyles = makeStyles({
     root: {
@@ -43,7 +41,7 @@ const useStyles = makeStyles({
                 Please note that the hours may change for any reason at any time.
               </p>
               <Paper className={classes.root}>
-              <table className={classes.table} aria-label="spanning table">
+              <Table className={classes.table} aria-label="spanning table">
               <TableHead>
                     <TableRow>
                       <TableCell><b>Facility</b></TableCell>
@@ -97,7 +95,7 @@ const useStyles = makeStyles({
                     <TableCell>Dawn - Dusk</TableCell>
                   </TableRow>
                 </TableBody>
-              </table> 
+              </Table> 
               </Paper>
               <a><i>
                   *If you are on the outdoor facilities during off-hours, police may be called and you will
